GATEWAY-2810 Updates for PD Orchestration
