+++ Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/patientdiscovery/entity/AbstractOutboundPatientDiscoveryOrchestrationContextBuilder.java
@@ -1,0 +1,136 @@
+/*
+ * Copyright (c) 2012, United States Government, as represented by the Secretary of Health and Human Services. 
+ * All rights reserved. 
+ *
+ * Redistribution and use in source and binary forms, with or without 
+ * modification, are permitted provided that the following conditions are met: 
+ *     * Redistributions of source code must retain the above 
+ *       copyright notice, this list of conditions and the following disclaimer. 
+ *     * Redistributions in binary form must reproduce the above copyright 
+ *       notice, this list of conditions and the following disclaimer in the documentation 
+ *       and/or other materials provided with the distribution. 
+ *     * Neither the name of the United States Government nor the 
+ *       names of its contributors may be used to endorse or promote products 
+ *       derived from this software without specific prior written permission. 
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
+ * DISCLAIMED. IN NO EVENT SHALL THE UNITED STATES GOVERNMENT BE LIABLE FOR ANY 
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES 
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; 
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND 
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT 
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS 
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
+ */
+package gov.hhs.fha.nhinc.patientdiscovery.entity;
+
+import org.hl7.v3.PRPAIN201305UV02;
+
+import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
+import gov.hhs.fha.nhinc.common.nhinccommon.NhinTargetSystemType;
+import gov.hhs.fha.nhinc.orchestration.AuditTransformer;
+import gov.hhs.fha.nhinc.orchestration.OutboundDelegate;
+import gov.hhs.fha.nhinc.orchestration.OutboundResponseProcessor;
+import gov.hhs.fha.nhinc.orchestration.OrchestrationContext;
+import gov.hhs.fha.nhinc.orchestration.OrchestrationContextBuilder;
+import gov.hhs.fha.nhinc.orchestration.PolicyTransformer;
+
+/**
+ * @author bhumphrey/paul
+ * 
+ */
+public abstract class AbstractOutboundPatientDiscoveryOrchestrationContextBuilder implements OrchestrationContextBuilder {
+
+    private NhinTargetSystemType target = null;
+    private PRPAIN201305UV02 request = null;
+    private AssertionType assertionType = null;
+    private PolicyTransformer policyTransformer = null;
+    private AuditTransformer auditTransformer = null;
+    private OutboundDelegate nhinDelegate = null;
+    private OutboundResponseProcessor nhinProcessor = null;
+    private String serviceName = "";
+
+    @Override
+    public OrchestrationContext build() {
+        return new OrchestrationContext(getStrategy(), getOrchestratable());
+    }
+
+    abstract protected OutboundPatientDiscoveryOrchestratable getOrchestratable();
+
+    abstract protected OutboundPatientDiscoveryStrategy getStrategy();
+
+    public void setTarget(NhinTargetSystemType t) {
+        this.target = t;
+    }
+
+    protected NhinTargetSystemType getTargetSystemType() {
+        return this.target;
+    }
+
+    public void setRequest(PRPAIN201305UV02 pdRequest) {
+        this.request = pdRequest;
+    }
+
+    protected PRPAIN201305UV02 getRequest() {
+        return this.request;
+    }
+
+    protected AssertionType getAssertionType() {
+        return assertionType;
+    }
+
+    public void setAssertionType(AssertionType assertionType) {
+        this.assertionType = assertionType;
+    }
+
+    protected PolicyTransformer getPolicyTransformer() {
+        return policyTransformer;
+    }
+
+    public void setPolicyTransformer(PolicyTransformer policyTransformer) {
+        this.policyTransformer = policyTransformer;
+    }
+
+    protected AuditTransformer getAuditTransformer() {
+        return auditTransformer;
+    }
+
+    public void setAuditTransformer(AuditTransformer auditTransformer) {
+        this.auditTransformer = auditTransformer;
+    }
+
+    protected OutboundDelegate getNhinDelegate() {
+        return nhinDelegate;
+    }
+
+    public void setNhinDelegate(OutboundDelegate nhinDelegate) {
+        this.nhinDelegate = nhinDelegate;
+    }
+
+    protected OutboundResponseProcessor getAggregator() {
+        return nhinProcessor;
+    }
+
+    public void setAggregator(OutboundResponseProcessor processor) {
+        nhinProcessor = processor;
+    }
+
+    protected OutboundResponseProcessor getProcessor() {
+        return nhinProcessor;
+    }
+
+    public void setProcessor(OutboundResponseProcessor processor) {
+        this.nhinProcessor = processor;
+    }
+
+    protected String getServiceName() {
+        return serviceName;
+    }
+
+    public void setServiceName(String name) {
+        this.serviceName = name;
+    }
+
+}
