+++ Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/patientdiscovery/entity/OutboundPatientDiscoveryDelegate.java
@@ -74,7 +74,7 @@ public class OutboundPatientDiscoveryDelegate implements OutboundDelegate {
     public OutboundPatientDiscoveryOrchestratable process(OutboundPatientDiscoveryOrchestratable message) {
         getLogger().debug("NhinPatientDiscoveryDelegate::process EntityPatientDiscoveryOrchestratable");
 
-        OutboundPatientDiscoveryOrchestrationContextBuilder contextBuilder = (OutboundPatientDiscoveryOrchestrationContextBuilder) OrchestrationContextFactory
+        AbstractOutboundPatientDiscoveryOrchestrationContextBuilder contextBuilder = (AbstractOutboundPatientDiscoveryOrchestrationContextBuilder) OrchestrationContextFactory
                 .getInstance().getBuilder(message.getTarget().getHomeCommunity(), NhincConstants.NHIN_SERVICE_NAMES.PATIENT_DISCOVERY);
 
         contextBuilder.setAssertionType(message.getAssertion());
