+++ Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/patientdiscovery/entity/OutboundPatientDiscoveryFactory.java
@@ -49,11 +49,9 @@ public class OutboundPatientDiscoveryFactory {
     public OrchestrationContextBuilder createOrchestrationContextBuilder(NhincConstants.GATEWAY_API_LEVEL apiLevel) {
         log.debug("EntityPatientDiscoveryFactory has apiLevel=" + apiLevel.toString());
         switch (apiLevel) {
-        case LEVEL_g0:
-            return new OutboundPatientDiscoveryOrchestrationContextBuilder_g0();        
         default:
-            log.warn("Unexpected api level received " + apiLevel.toString() + ".  Defaulting to g0");
-            return new OutboundPatientDiscoveryOrchestrationContextBuilder_g0();
+            log.warn("The same implementation is used for both PD g0 and PD g1.");
+            return new OutboundPatientDiscoveryOrchestrationContextBuilder();
         }
     }
